pyproject.toml
[tool.poetry]
name = "cryptocurrency-rate"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
websockets = "^12.0"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
python-binance = "^1.0.19"
pika = "^1.3.2"
aio-pika = "^9.3.1"
pycoingecko = "^3.1.0"
aiohttp = "^3.9.1"
python-dotenv = "^1.0.0"
sqlalchemy = {extras = ["async"], version = "^2.0.23"}
pydantic = {extras = ["dotenv"], version = "^2.5.2"}
asyncpg = "^0.29.0"
alembic = "^1.12.1"
pydantic-settings = "^2.1.0"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pre-commit = "^3.5.0"
flake8 = "^6.1.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
